+++ src/main/java/blackjack/Application.java
@@ -5,11 +5,11 @@ import static java.util.function.Function.identity;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
 
-import blackjack.domain.ProfitCalculator;
+import blackjack.domain.betting.ProfitCalculator;
 import blackjack.domain.betting.BettingAmount;
 import blackjack.domain.gambler.Names;
-import blackjack.domain.Round;
-import blackjack.domain.WinningDiscriminator;
+import blackjack.domain.game.Round;
+import blackjack.domain.game.WinningDiscriminator;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardDeck;
 import blackjack.domain.card.CardShape;
@@ -18,7 +18,7 @@ import blackjack.domain.card.CardType;
 import blackjack.domain.gambler.Name;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
-import blackjack.view.WinningType;
+import blackjack.domain.game.WinningType;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
