+++ src/main/java/blackjack/domain/betting/ProfitCalculator.java
@@ -1,8 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.betting;
 
-import blackjack.domain.betting.BettingAmount;
 import blackjack.domain.gambler.Name;
-import blackjack.view.WinningType;
+import blackjack.domain.game.WinningType;
 import java.util.Map;
 
 public class ProfitCalculator {
