+++ src/main/java/blackjack/domain/game/WinningType.java
@@ -1,4 +1,4 @@
-package blackjack.view;
+package blackjack.domain.game;
 
 import static java.util.function.Function.identity;
 import static java.util.stream.Collectors.toMap;
