+++ src/test/java/blackjack/view/WinningTypeTest.java
@@ -2,6 +2,7 @@ package blackjack.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.game.WinningType;
 import blackjack.domain.betting.BettingAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
